+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - [Changed the way of adding new column in Table Input Widget][11388]. The
   "virtual column" is replaced with an explicit (+) button.
 - [New dropdown-based component menu][11398].
+- [Methods defined on Standard.Base.Any type are now visible on all
+  components][11451].
 - [Undo/redo buttons in the top bar][11433].
 - [Size of Table Input Widget is preserved and restored after project
   re-opening][11435]
@@ -28,6 +30,7 @@
 [11383]: https://github.com/enso-org/enso/pull/11383
 [11388]: https://github.com/enso-org/enso/pull/11388
 [11398]: https://github.com/enso-org/enso/pull/11398
+[11451]: https://github.com/enso-org/enso/pull/11451
 [11433]: https://github.com/enso-org/enso/pull/11433
 [11435]: https://github.com/enso-org/enso/pull/11435
 [11446]: https://github.com/enso-org/enso/pull/11446